% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{geom_hurricane}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{A Dataframe Object containing the data to plot.}

\item{x}{Column of `data` depicting the x-locations of the Storm's Center (LONGITUDE for maps).}

\item{y}{Column of `data` depicting the y-locations of the Storm's Center (LATITUDE for maps).}

\item{r_ne}{Column of `data` depicting the wind_raius of the North East Quadrant.}

\item{r_se}{Column of `data` depicting the wind_raius of the South East Quadrant.}

\item{r_nw}{Column of `data` depicting the wind_raius of the North West Quadrant.}

\item{r_sw}{Column of `data` depicting the wind_raius of the South West Quadrant.}

\item{fill}{OPTIONAL. Column of `data` which will be use to color the circles.}

\item{color}{OPTIONAL. Column of `data` which will be use to color the circle outlines.}

\item{scale_radii}{OPTIONAL. Numeric. Size scale of the drawn Hurricane.}
}
\value{
This function adds the hurricane image (Grob) into the current graphics device.
}
\description{
This function return a Grob object to be displayed. The image will characterize a hurricane depited by the edges of each circle's
quadrant (possibly representing wind_raius), and the color of the circles (possibly depicting wind_speeds).
}
\examples{

\dontrun{

## Data
ext_tracks <- reading_tropical_cyclone(filename = "data/ebtrk_atlc_1988_2015.txt")
ext_tracks <- tc_geom_format(ext_tracks)

## STORMS

katrina <- ext_tracks \%>\%
    dplyr::filter(storm_id == "KATRINA-2005", date == "2005-08-29 12:00:00") \%>\%
    dplyr::mutate(longitude = longitude * -1) \%>\%
    as.data.frame()

## Maps

ggmap::get_map("Louisiana", zoom = 6, maptype = "toner-background") \%>\%
    ggmap::ggmap(extent = "device") +
    geom_hurricane(data = katrina,
                   aes(x = longitude, y = latitude,
                       r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
                       fill = wind_speed, color = wind_speed)) +
    ggplot2::scale_color_manual(name = "Wind speed (kts)",
                               values = c("red", "orange", "yellow")) +
    ggplot2::scale_fill_manual(name = "Wind speed (kts)",
                               values = c("red", "orange", "yellow"))

}

}
