% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomHurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{GeomHurricane}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\value{
This function returns a Data Frame object, with row names given based on the \code{row_names} argument.
        It includes all the columns of the given dataset whose elements are all numeric or have
        a meaningful numerical equivalent.
}
\description{
This Geom returns ggproto object made of a gTree of polygonGrobs. Each Grob is a circle with different radii in each circle depending on the quadrant.
The grobs are colored based on the \code{colour} argument. There will be one grob per wind_speed, aka \code{colour} variable. This
is a ggproto object, and thus it requires no Arguments once Created (those are found in the calling function).
}
\section{Errors}{
 \code{GeomHurricane} won't be able to display the Hurricane grobs if there isn't a current ggmap
         active for which this Geom will be apllied onto.
}

\examples{

\dontrun{
## Example of Creating a calling function with this Grob
geom_hurricane <- function(mapping = NULL, data = NULL, stat = "identity",
                           position = "identity", na.rm = FALSE,
                           show.legend = NA, inherit.aes = TRUE, ...) {
                                 ggplot2::layer(geom = GeomHurricane, mapping = mapping,
                                                data = data, stat = stat, position = position,
                                                show.legend = show.legend, inherit.aes = inherit.aes,
                                                params = list(na.rm = na.rm, ...))
                           }

}

}
\keyword{datasets}
