% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_tropical_cyclone.R
\name{reading_tropical_cyclone}
\alias{reading_tropical_cyclone}
\title{reading_tropical_cyclone}
\usage{
reading_tropical_cyclone(filename)
}
\arguments{
\item{filename}{Either a path to a file, or a connection.
Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded.}
}
\value{
This function returns the raw Tropical Cyclone Data, in \code{tbl_df} format.
}
\description{
This function correctly reads the The Tropical Cyclone Extended Best Track Dataset.
 **NOTE: This function was Created by the team behind COursera's Building Data Visualization Tools course.
}
\examples{

\dontrun{

ext_tracks <-reading_tropical_cyclone(filename = "data/ebtrk_atlc_1988_2015.txt")

dataset_url <- "http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/data/ebtrk_atlc_1988_2015.txt"
ext_tracks <- reading_tropical_cyclone(filename = dataset_url)

}

}
