% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_geom_format.R
\name{tc_geom_format}
\alias{tc_geom_format}
\title{tc_geom_format}
\usage{
tc_geom_format(dataset)
}
\arguments{
\item{dataset}{A dataframe object. Must be read using the \code{reading_tropical_cyclone} function, or share the same format
as data read that way.}
}
\value{
This function returns a Data Frame object of The Tropical Cyclone Extended Best Track Dataset, formatted
        down to N rows per Storm Observation (Where N represents each available wind_speed).
}
\description{
This function format, subsets, and mutates The Tropical Cyclone Extended Best Track Dataset that has been read
using the \code{reading_tropical_cyclone} function. It returns one row per wind_speed, per Storm Observation.
Individual Storm observations ae characterized by their wind_radius (per direction), and storm metadata.
}
\examples{

\dontrun{

## Reading Dataset
data_tc <- reading_tropical_cyclone(filename = "data/ebtrk_atlc_1988_2015.txt")
ext_tracks <- tc_geom_format(data_tc) # Formatting

}

}
